pull the `Kinded` trait into `chalk::lowering`
This severs the unfortunate relationship between `chalk_ir` and
`chalk_parse`.
